+++ entity/cattlebirthscreen.cpp
