+++ entity/cattlebirthscreen.h
