+++ entity/cattlebuyscreen.cpp
