+++ entity/cattlemanagementscreen.cpp
