+++ entity/cattleremovescreen.cpp
