+++ entity/cattleremovescreen.h
