+++ entity/financialmanagementscreen.cpp
