+++ entity/financialrecordscreen.cpp
