+++ entity/financialremovescreen.cpp
