+++ lib/cattleImpl.cpp
