--- lib/src.cpp
@@ -1,7 +1,0 @@
-#include "stdio.h"
-
-int m(){
-    printf("Hello World!");
-
-    return 0;
-}
