+++ views/cattlebirthscreen.ui
