+++ views/cattlebuyscreen.ui
