+++ views/cattlemanagementscreen.ui
