+++ views/cattleregisterscreen.ui
