+++ views/cattleremovescreen.ui
