+++ views/financialmanagementscreen.ui
