+++ views/financialrecordscreen.ui
